"interface-over-type-literal": false
Justification: there are many times (and @XVincentX would probably agree with me) when it makes more sense to export a type alias rather than an interface. Semantically, I think of an `interface` as a thing meant to be implemented / extended. If I'm just documenting a return type of a function (or a postMessage message format), like `{ result: string, err?: { code: number, msg: string} }`, it feels wrong to declare it as an "interface".